requirements.txt
py-cord>=2.0.0
jikanpy-v4
python-dotenv